extract document migrator into multiple files
